Create README.md
Threaded Python OpenCV camera stream with FPS measurement
This is my first GitHub contribution.
It´s a Python Module than give threaded camera stream with the capability to get the camera reading FPS 
If you run the script it have a little test example. 
You could use it as a imported module.